Moved stuff to radio_interface file
